pkg/seq/measure: improves file organization
